Bitchin' -- I fixed the CI build
The activator script is using SBT_OPTIONS, but Travis CI is setting that
to a path (with a preceeding `@` symbol, for some reason). Changing this
to the expected sequence of JVM options fixes the build.

Email notifications are now enabled.

Also added a build status image to the readme, so the status of the
build can be determined from the GitHub page.