Integrated Duan's model
Renamed some attributes of the case classes. It's mostly as issue #7
described, although we couldn't have `type` in Inspection, since that's
a reserved word. I also realize that the IDs of locations is something
we might need to know. We also had missed the RHAs. I removed the
province from Duan's model, as that's always the same.