Created scripts to start/stop the server
Note: these are intended only to be run on the production server.

Usage: Just run `./startServer.sh` and that should be sufficient. It'll
try and stop the existing server, if one is running (to allow for
restarting the server). It simply runs Play in production mode with
nohup to allow you to hangup the SSH connection without the program
exiting (obviously desirable for a daemon).

This script must be manually run and will have to be run if this server
ever restarts (not clear if we should make it automatically start).

Starting the server creates 2 files: nohup.out (normal server output,
mostly uneeded) and errors.log (the output to stderr). Play also
maintains a file storing the PID of the server (see the source of
stopServer.sh). If Play fails to shut down properly, that file might
have to be manually removed (as new instances will not start with it
existing).