Fixing pixel again
This required making changes to the bootstrap stylesheet file directly.
This is suboptimal, but the template engine refuses to load our main.less
file with a higher priority than the bootstrap.min.css file, so the
only way to have a long term fix is to go straight to the source.