Find location can now search
Searching is done iff an item from the dropdown is not chosen. Hints
have been disabled because they complicate the search heavily.

Issue with hints is that they populate the `obj` parameter of
`onSubmit`, and can't be differentiated from an `obj` that was chosen by
selecting from the typeahead results. This is bad because the hint is
only the name, and might match multiple locations (who knows which the
user intended?). If a location was chosen from the results, this is not
an issue because the results display the address.