Using custom dictionary for synonyms
This allows us to provide synonyms applicable to our domain. For now,
I've populated it merely with road abbreviations so that users can
search for "8th street" and get results for "8th st" (etc).

The synonyms file must be placed in /usr/share/postgresql/9.3/tsearch_data
(where 9.3 is the PostgreSQL version). Otherwise you'll get an error
about not being able to find that file (the error message will specify
the directory that was searched in, so it should be easy to resolve).