Start to the admin interface
Listing all locations in a table with filtering, sorting, etc. Links for
editing and deleting locations still to come.

Note: This part of the site has not (yet?) been designed with mobile in
mind. This is because of the tables not easily fitting in. Seems like a
reasonable limitation since the admin interface isn't something that
regular users access.